jobs/colorize/requirements.txt
google-cloud-storage
google-cloud-firestore
google-genai
functions-framework
python-dotenv
Pillow
tenacity